requirements.txt
python-magic>=0.4.27
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.0.0
PyPDF2>=3.0.0
simhash>=2.1.2
python-Levenshtein>=0.21.0
watchdog>=3.0.0
PyYAML>=6.0
extract-msg>=0.45.0
imapclient>=2.3.0
tqdm>=4.65.0
